pyproject.toml
[tool.poetry]
name = "detectionmetrics"
version = "2.1.0"
description = "Tools for evaluating segmentation and object detection models"
authors = ["JdeRobot", "d.pascualhe <d.pascualhe@gmail.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.67.0"
pandas = "^2.2.3"
PyYAML = "^6.0.2"
pyarrow = "^18.0.0"
pillow = "^11.0.0"
numpy = "1.26.4"
opencv-python-headless = "^4.10.0.84"
scikit-learn = "^1.6.0"
open3d = "^0.18.0"
addict = "^2.4.0"
matplotlib = "^3.10.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
